Relocate CLI error helpers to shared module
Move scripts/utils/cli-errors.js into src/shared/cli/cli-errors.js so the Prettier wrapper no longer depends on script-only utilities while generator scripts keep reusing the shared error helpers.
