+++ src/shared/cli/cli-errors.js
